+++ src/main/java/net/frozenblock/wilderwild/client/WWParticleEngine.java
@@ -31,6 +31,7 @@ import net.frozenblock.wilderwild.particle.MesogleaBubbleParticle;
 import net.frozenblock.wilderwild.particle.MesogleaBubblePopParticle;
 import net.frozenblock.wilderwild.particle.MesogleaCurrentDownParticle;
 import net.frozenblock.wilderwild.particle.MesogleaDripParticle;
+import net.frozenblock.wilderwild.particle.MesogleaSplashParticle;
 import net.frozenblock.wilderwild.particle.PollenParticle;
 import net.frozenblock.wilderwild.particle.SeedParticle;
 import net.frozenblock.wilderwild.particle.ShriekerBubbleSeedParticle;
@@ -38,7 +39,6 @@ import net.frozenblock.wilderwild.particle.TermiteParticle;
 import net.frozenblock.wilderwild.particle.WindParticle;
 import net.frozenblock.wilderwild.particle.factory.WilderParticleFactories;
 import net.frozenblock.wilderwild.registry.WWParticleTypes;
-import net.minecraft.client.particle.SplashParticle;
 
 @Environment(EnvType.CLIENT)
 public final class WWParticleEngine {
@@ -90,7 +90,7 @@ public final class WWParticleEngine {
 			(spriteProvider) -> new MesogleaCurrentDownParticle.BubbleFactory(spriteProvider, WWParticleTypes.BLUE_PEARLESCENT_MESOGLEA_BUBBLE_POP)
 		);
 		particleRegistry.register(WWParticleTypes.BLUE_PEARLESCENT_MESOGLEA_BUBBLE_POP, MesogleaBubblePopParticle.BubbleFactory::new);
-		particleRegistry.register(WWParticleTypes.BLUE_PEARLESCENT_MESOGLEA_SPLASH, SplashParticle.Provider::new);
+		particleRegistry.register(WWParticleTypes.BLUE_PEARLESCENT_MESOGLEA_SPLASH, MesogleaSplashParticle.Provider::new);
 		particleRegistry.register(WWParticleTypes.PURPLE_PEARLESCENT_HANGING_MESOGLEA, MesogleaDripParticle.PPMesogleaHangProvider::new);
 		particleRegistry.register(WWParticleTypes.PURPLE_PEARLESCENT_FALLING_MESOGLEA, MesogleaDripParticle.PPMesogleaFallProvider::new);
 		particleRegistry.register(WWParticleTypes.PURPLE_PEARLESCENT_LANDING_MESOGLEA, MesogleaDripParticle.PPMesogleaLandProvider::new);
@@ -107,7 +107,7 @@ public final class WWParticleEngine {
 			(spriteProvider) -> new MesogleaCurrentDownParticle.BubbleFactory(spriteProvider, WWParticleTypes.PURPLE_PEARLESCENT_MESOGLEA_BUBBLE_POP)
 		);
 		particleRegistry.register(WWParticleTypes.PURPLE_PEARLESCENT_MESOGLEA_BUBBLE_POP, MesogleaBubblePopParticle.BubbleFactory::new);
-		particleRegistry.register(WWParticleTypes.PURPLE_PEARLESCENT_MESOGLEA_SPLASH, SplashParticle.Provider::new);
+		particleRegistry.register(WWParticleTypes.PURPLE_PEARLESCENT_MESOGLEA_SPLASH, MesogleaSplashParticle.Provider::new);
 		particleRegistry.register(WWParticleTypes.BLUE_HANGING_MESOGLEA, MesogleaDripParticle.BMesogleaHangProvider::new);
 		particleRegistry.register(WWParticleTypes.BLUE_FALLING_MESOGLEA, MesogleaDripParticle.BMesogleaFallProvider::new);
 		particleRegistry.register(WWParticleTypes.BLUE_LANDING_MESOGLEA, MesogleaDripParticle.BMesogleaLandProvider::new);
@@ -124,7 +124,7 @@ public final class WWParticleEngine {
 			(spriteProvider) -> new MesogleaCurrentDownParticle.BubbleFactory(spriteProvider, WWParticleTypes.BLUE_MESOGLEA_BUBBLE_POP)
 		);
 		particleRegistry.register(WWParticleTypes.BLUE_MESOGLEA_BUBBLE_POP, MesogleaBubblePopParticle.BubbleFactory::new);
-		particleRegistry.register(WWParticleTypes.BLUE_MESOGLEA_SPLASH, SplashParticle.Provider::new);
+		particleRegistry.register(WWParticleTypes.BLUE_MESOGLEA_SPLASH, MesogleaSplashParticle.Provider::new);
 		particleRegistry.register(WWParticleTypes.YELLOW_HANGING_MESOGLEA, MesogleaDripParticle.YMesogleaHangProvider::new);
 		particleRegistry.register(WWParticleTypes.YELLOW_FALLING_MESOGLEA, MesogleaDripParticle.YMesogleaFallProvider::new);
 		particleRegistry.register(WWParticleTypes.YELLOW_LANDING_MESOGLEA, MesogleaDripParticle.YMesogleaLandProvider::new);
@@ -141,7 +141,7 @@ public final class WWParticleEngine {
 			(spriteProvider) -> new MesogleaCurrentDownParticle.BubbleFactory(spriteProvider, WWParticleTypes.YELLOW_MESOGLEA_BUBBLE_POP)
 		);
 		particleRegistry.register(WWParticleTypes.YELLOW_MESOGLEA_BUBBLE_POP, MesogleaBubblePopParticle.BubbleFactory::new);
-		particleRegistry.register(WWParticleTypes.YELLOW_MESOGLEA_SPLASH, SplashParticle.Provider::new);
+		particleRegistry.register(WWParticleTypes.YELLOW_MESOGLEA_SPLASH, MesogleaSplashParticle.Provider::new);
 		particleRegistry.register(WWParticleTypes.LIME_HANGING_MESOGLEA, MesogleaDripParticle.LMesogleaHangProvider::new);
 		particleRegistry.register(WWParticleTypes.LIME_FALLING_MESOGLEA, MesogleaDripParticle.LMesogleaFallProvider::new);
 		particleRegistry.register(WWParticleTypes.LIME_LANDING_MESOGLEA, MesogleaDripParticle.LMesogleaLandProvider::new);
@@ -158,7 +158,7 @@ public final class WWParticleEngine {
 			(spriteProvider) -> new MesogleaCurrentDownParticle.BubbleFactory(spriteProvider, WWParticleTypes.LIME_MESOGLEA_BUBBLE_POP)
 		);
 		particleRegistry.register(WWParticleTypes.LIME_MESOGLEA_BUBBLE_POP, MesogleaBubblePopParticle.BubbleFactory::new);
-		particleRegistry.register(WWParticleTypes.LIME_MESOGLEA_SPLASH, SplashParticle.Provider::new);
+		particleRegistry.register(WWParticleTypes.LIME_MESOGLEA_SPLASH, MesogleaSplashParticle.Provider::new);
 		particleRegistry.register(WWParticleTypes.PINK_HANGING_MESOGLEA, MesogleaDripParticle.PMesogleaHangProvider::new);
 		particleRegistry.register(WWParticleTypes.PINK_FALLING_MESOGLEA, MesogleaDripParticle.PMesogleaFallProvider::new);
 		particleRegistry.register(WWParticleTypes.PINK_LANDING_MESOGLEA, MesogleaDripParticle.PMesogleaLandProvider::new);
@@ -175,7 +175,7 @@ public final class WWParticleEngine {
 			(spriteProvider) -> new MesogleaCurrentDownParticle.BubbleFactory(spriteProvider, WWParticleTypes.PINK_MESOGLEA_BUBBLE_POP)
 		);
 		particleRegistry.register(WWParticleTypes.PINK_MESOGLEA_BUBBLE_POP, MesogleaBubblePopParticle.BubbleFactory::new);
-		particleRegistry.register(WWParticleTypes.PINK_MESOGLEA_SPLASH, SplashParticle.Provider::new);
+		particleRegistry.register(WWParticleTypes.PINK_MESOGLEA_SPLASH, MesogleaSplashParticle.Provider::new);
 		particleRegistry.register(WWParticleTypes.RED_HANGING_MESOGLEA, MesogleaDripParticle.RMesogleaHangProvider::new);
 		particleRegistry.register(WWParticleTypes.RED_FALLING_MESOGLEA, MesogleaDripParticle.RMesogleaFallProvider::new);
 		particleRegistry.register(WWParticleTypes.RED_LANDING_MESOGLEA, MesogleaDripParticle.RMesogleaLandProvider::new);
@@ -192,6 +192,6 @@ public final class WWParticleEngine {
 			(spriteProvider) -> new MesogleaCurrentDownParticle.BubbleFactory(spriteProvider, WWParticleTypes.RED_MESOGLEA_BUBBLE_POP)
 		);
 		particleRegistry.register(WWParticleTypes.RED_MESOGLEA_BUBBLE_POP, MesogleaBubblePopParticle.BubbleFactory::new);
-		particleRegistry.register(WWParticleTypes.RED_MESOGLEA_SPLASH, SplashParticle.Provider::new);
+		particleRegistry.register(WWParticleTypes.RED_MESOGLEA_SPLASH, MesogleaSplashParticle.Provider::new);
 	}
 }